Move Sniffer to different folder
