+++ pkg/transport/network/sniffer/decoder_test.go
@@ -1,4 +1,4 @@
-package network
+package sniffer
 
 import (
 	"log"
